Reorganized dao responses structure
