+++ src/main/java/com/battlemetrics/service/MailContentBuilder.java
@@ -1,7 +1,7 @@
 package com.battlemetrics.service;
 
-import com.battlemetrics.dao.response.PlayerSessionResponse;
-import com.battlemetrics.dao.response.ServerResponse;
+import com.battlemetrics.dao.response.bmapi.PlayerSessionResponse;
+import com.battlemetrics.dao.response.bmapi.ServerResponse;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
